Rename context file for export.
